Make compressed reads interrupt-safe
A read operation might get an "interrupted" error, in which case the
correct behavior is (usually) to try again as if nothing happened.

When we read a compressed stream we check whether the underlying
reader received an error. But we should only check whether it received
an error for the latest read, so that we can ignore these interrupts
properly.

This is the only place I noticed where interrupts were handled
improperly.

AFAIK this can't happen in reality because we don't install signal
handlers, but it's good practice.